Fix check of pip cache

+++ pip-bundler.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python2
 
+import cgi
 import pip
 import pkg_resources
 import sys, os, tarfile, zipfile, shutil, urllib2, getopt
@@ -21,7 +22,12 @@ if StrictVersion(pip.__version__) < StrictVersion('1.4.1'):
 class Bundler:
     def __init__(self, cache_directory, use_wheel=False):
         self.cache_directory = cache_directory
-        self.finder = PackageFinder([], ["http://pypi.python.org/simple"], use_wheel=use_wheel)
+        self.local_finder = PackageFinder([self.cache_directory],
+                                          [],
+                                          use_wheel=use_wheel)
+        self.external_finder = PackageFinder([],
+                                             ["http://pypi.python.org/simple"],
+                                             use_wheel=use_wheel)
 
     def _is_cached(self, file_name):
         return os.path.isfile(os.path.join(self.cache_directory, file_name))
@@ -29,7 +35,8 @@ class Bundler:
     def _fetch_package(self,package):
         print(" -> downloading %s" % package.filename)
         remote_file = urllib2.urlopen(package.url)
-        local_file = open(str(os.path.join(self.cache_directory, package.filename)), 'wb')
+        local_file = open(str(os.path.join(self.cache_directory,
+                                           package.filename)), 'wb')
         local_file.write(remote_file.read())
         local_file.close()
 
@@ -39,9 +46,11 @@ class Bundler:
         dependencies = []
 
         if file_name.find('.tar.gz') > 0 or file_name.find('.tar.bz2') > 0:
-            archive = tarfile.TarFile.open(os.path.join(self.cache_directory, file_name))
+            archive = tarfile.TarFile.open(os.path.join(self.cache_directory,
+                                                        file_name))
         elif file_name.find('.zip') > 0:
-            archive = zipfile.ZipFile(os.path.join(self.cache_directory, file_name))
+            archive = zipfile.ZipFile(os.path.join(self.cache_directory,
+                                                   file_name))
 
         if type(archive) is zipfile.ZipFile:
             names = archive.namelist()
@@ -53,11 +62,15 @@ class Bundler:
             package_name = str(splitext(file_name)[0]).strip()
 
             if os.path.join(package_name, "requirements.txt") in names:
-                archive.extract(os.path.join(package_name, "requirements.txt"), "tmp")
-                package_requires = os.path.join("tmp", package_name, "requirements.txt")
+                archive.extract(os.path.join(package_name, "requirements.txt"),
+                                "tmp")
+                package_requires = os.path.join("tmp", package_name,
+                                                "requirements.txt")
             elif os.path.join(package_name, "tools/pip-requires") in names:
-                archive.extract(os.path.join(package_name, "tools/pip-requires"), "tmp")
-                package_requires = os.path.join("tmp", package_name, "tools/pip-requires")
+                archive.extract(os.path.join(package_name,
+                                             "tools/pip-requires"), "tmp")
+                package_requires = os.path.join("tmp", package_name,
+                                                "tools/pip-requires")
 
             if package_requires is not None:
                 install_reqs = parse_requirements(package_requires)
@@ -73,18 +86,26 @@ class Bundler:
 
     def resolve(self, requires, parent=None, update=False):
         self.update = update
+
+        if self.update:
+            finder = self.external_finder
+        else:
+            finder = self.local_finder
+
         for require in requires:
             try:
-                package = self.finder.find_requirement(InstallRequirement.from_line(require, None), False)
+                package = finder.find_requirement(
+                    InstallRequirement.from_line(require, None), False)
                 if not self._is_cached(package.filename):
                     if self.update:
                         self._fetch_package(package)
                     else:
-                        print(" -> %s not cached but needed" % package.filename)
+                        print("-> %s not cached but needed" % package.filename)
                         exit(1)
 
                 dependencies = self._extract_dependencies(package.filename)
-                self.resolve(dependencies, update=self.update, parent=package.filename)
+                self.resolve(dependencies, update=self.update,
+                             parent=package.filename)
 
             except DistributionNotFound as error:
                 print(error)
@@ -125,6 +146,42 @@ class Dir2Pi:
             raise ValueError(msg)
         return split[0], pkg_resources.safe_name(split[1])
 
+    def create_indexes(self):
+        pkgdir = self.cache_directory
+        if not os.path.isdir(pkgdir):
+            raise ValueError("no such directory: %r" % (pkgdir, ))
+
+        pkgdirpath = lambda *x: os.path.join(pkgdir, *x)
+
+        if not os.path.isdir(pkgdirpath("simple")):
+            return
+
+        pkg_index = ("<html><head><title>Simple Index</title>"
+                     "<meta name='api-version' value='2' /></head><body>\n")
+
+        for file in os.listdir(pkgdir):
+            pkg_filepath = os.path.join(pkgdir, file)
+            if not os.path.isfile(pkg_filepath):
+                continue
+            pkg_basename = os.path.basename(file)
+            if pkg_basename.startswith("."):
+                continue 
+            pkg_name, pkg_rest = self.file_to_package(pkg_basename, pkgdir)
+            pkg_dir = pkgdirpath("simple", pkg_name)
+            if not os.path.exists(pkg_dir):
+                continue
+            pkg_new_basename = "-".join([pkg_name, pkg_rest])
+            pkg_name_html = cgi.escape(pkg_name)
+            pkg_index += "<a href='{0}/'>{0}</a><br />\n".format(pkg_name_html)
+            with open(os.path.join(pkg_dir, "index.html"), "a") as fp:
+                pkg_new_basename_html = cgi.escape(pkg_new_basename)
+                fp.write("<a href='%s'>%s</a><br />\n"
+                         %(pkg_new_basename_html, pkg_new_basename_html))
+
+        with open(pkgdirpath("simple/index.html"), "w") as fp:
+            fp.write(pkg_index)
+
+
     def create_dir_simple(self):
         """
             Creates the directory self.cache_directory/simple/ and populates
@@ -151,6 +208,7 @@ class Dir2Pi:
         pkgdir = self.cache_directory
         if not os.path.isdir(pkgdir):
             raise ValueError("no such directory: %r" % (pkgdir, ))
+        
         pkgdirpath = lambda *x: os.path.join(pkgdir, *x)
 
         shutil.rmtree(pkgdirpath("simple"), ignore_errors=True)
@@ -169,9 +227,9 @@ class Dir2Pi:
                 os.mkdir(pkg_dir)
             pkg_new_basename = "-".join([pkg_name, pkg_rest])
             symlink_target = os.path.join(pkg_dir, pkg_new_basename)
-            symlink_source = os.path.join("../../", pkg_basename)
-            os.symlink(symlink_source, symlink_target)
-
+            if not os.path.exists(symlink_target):
+                symlink_source = os.path.join("../../", pkg_basename)
+                os.symlink(symlink_source, symlink_target)
 
 def print_help():
     print 'pip-bundler.py <options> -c <cache directory> requires'
@@ -198,11 +256,15 @@ def main(argv):
     if cache_directory is None:
         print("Cache directory not configured")
         exit(1)
+    
+    dir2pi = Dir2Pi(str(cache_directory))
+    # Create indexes for each package. It needs for correct working of Bundler
+    dir2pi.create_indexes()
 
+    #Download packages include packages from requirements
     bundler = Bundler(str(cache_directory))
     bundler.resolve(args, update=cache_update)
 
-    dir2pi = Dir2Pi(str(cache_directory))
     dir2pi.create_dir_simple()
